Extracts Structs from sql to go builder to transaction objects
