+++ internal/common/builder/sql_to_go_builder.go
@@ -32,7 +32,6 @@
 package builder
 
 import (
-	"database/sql"
 	"encoding/json"
 	"fmt"
 	"sync"
@@ -41,38 +40,6 @@ import (
 	jsoniter "github.com/json-iterator/go"
 )
 
-// AssetAdministrationShellDescriptorRow represents a single SQL result row
-// for an Asset Administration Shell (AAS) descriptor. It carries nullable
-// string/integer columns from the database and foreign-key references to
-// related records such as administrative information, display names, and
-// descriptions.
-type AssetAdministrationShellDescriptorRow struct {
-	DescID        int64
-	AssetKindStr  sql.NullString
-	AssetType     sql.NullString
-	GlobalAssetID sql.NullString
-	IDShort       sql.NullString
-	IDStr         string
-	AdminInfoID   sql.NullInt64
-	DisplayNameID sql.NullInt64
-	DescriptionID sql.NullInt64
-}
-
-// SubmodelDescriptorRow represents a single SQL result row for a Submodel
-// descriptor that is associated with an AAS descriptor. It includes the
-// database identifiers of the AAS and Submodel descriptors as well as
-// optional columns and foreign-key references such as semantic reference
-// and administrative information.
-type SubmodelDescriptorRow struct {
-	AasDescID     int64
-	SmdDescID     int64
-	IDShort       sql.NullString
-	ID            sql.NullString
-	SemanticRefID sql.NullInt64
-	AdminInfoID   sql.NullInt64
-	DescriptionID sql.NullInt64
-	DisplayNameID sql.NullInt64
-}
 // ParseReferredReferencesFromRows parses referred reference data from already unmarshalled ReferredReferenceRow objects.
 //
 // This function handles the complex case where references point to other references (referred references).
